requirements.txt
requests
tweepy==4.15.0
python-dotenv
openai==0.28.0

